+++ src/Pim/Bundle/DataGridBundle/Datasource/Orm/OrmDatasource.php
@@ -21,6 +21,41 @@ class OrmDatasource extends OroOrmDatasource
     const TYPE = 'pim_orm';
 
     /**
+     * @var string
+     */
+    const IS_FLEXIBLE_ENTITY_PATH = '[source][is_flexible]';
+
+    /**
+     * @var string
+     */
+    const ENTITY_PATH = '[source][entity]';
+
+    /**
+     * @var string
+     */
+    const DISPLAYED_ATTRIBUTES_PATH = '[source][displayed_attribute_ids]';
+
+    /**
+     * @var string
+     */
+    const DISPLAYED_LOCALE_PATH = '[source][locale_code]';
+
+    /**
+     * @var string
+     */
+    const USEABLE_ATTRIBUTES_PATH = '[source][attributes_configuration]';
+
+    /**
+     * @var boolean
+     */
+    protected $isFlexible = false;
+
+    /**
+     * @var string
+     */
+    protected $localeCode = null;
+
+    /**
      * {@inheritdoc}
      */
     public function process(DatagridInterface $grid, array $config)
@@ -38,6 +73,9 @@ class OrmDatasource extends OroOrmDatasource
             $this->qb = $repository->createQueryBuilder('o');
         }
 
+        $this->isFlexible = isset($config['is_flexible']) ? (bool) $config['is_flexible'] : false;
+        $this->localeCode = isset($config['locale_code']) ? $config['locale_code'] : null;
+
         $grid->setDatasource(clone $this);
     }
 
@@ -48,10 +86,32 @@ class OrmDatasource extends OroOrmDatasource
     {
         $query = $this->qb->getQuery();
 
-        $results = $query->execute();
-        $rows    = [];
-        foreach ($results as $result) {
-            $rows[] = new ResultRecord($result);
+        if ($this->isFlexible) {
+            $results = $query->getArrayResult();
+            $rows    = [];
+            foreach ($results as $result) {
+                $entityFields = $result[0];
+                unset($result[0]);
+                $otherFields = $result;
+                $result = $entityFields + $otherFields;
+                $values = $result['values'];
+                foreach ($values as $value) {
+                    $result[$value['attribute']['code']]= $value;
+                }
+                unset($result['values']);
+                $result['dataLocale']= $this->localeCode;
+
+                $rows[] = new ResultRecord($result);
+            }
+
+        } else {
+            $results = $query->execute();
+            $rows    = [];
+            foreach ($results as $result) {
+                $rows[] = new ResultRecord($result);
+            }
+
+            return $rows;
         }
 
         return $rows;