+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Uservoice integration
 - Add a last operations widget on the dashboard
 - Add new user as fixtures
+- Auto-refresh job execution report page
 
 ## Improvements
 - Load choices for grid filters asynchronously
@@ -15,6 +16,8 @@
 - Prevent hydrating all attributes in the available attributes addition form
 - Prevent hydrating all families in the product edition form
 - Import conversion units for channels
+- Product grid loading performance by hydrating as array and introduce selector extension
+- Add a screen to select the attribute type before creating an attribute
 
 ## Bug fixes
 - Fix non-updated values being displayed in the the audit history
@@ -24,6 +27,8 @@
 - Restored missing attributes translation
 - Fix the display of scopable metric attributes in the product edit form
 - Fix regression with resolve target entity on Category entity
+- Fix datepicker in date attribute form
+- Fix unwanted fields appearing in attribute creation form if server-side validation errors are present
 
 ## BC breaks
 - Remove the date type property of Attribute and simplify the pim_catalog_date attribute type to support date only (not date/datetime/time)
@@ -31,11 +36,14 @@
 - Rename findByWithSortedAttribute to findOneByWithValues, add pre-select attributes and related translations to reduce number of lazy loaded queries when edit a product
 - Rename findByWithAttributes to findAllByAttributes
 - MeasureBundle has been moved from the BAP to an external repository (akeneo/measure-bundle).
-- AkeneoBundle has been moved from the BAP to an external repository (akeneo/batch-bundle).
+- BatchBundle has been moved from the BAP to an external repository (akeneo/batch-bundle).
 - Remove magic setter access to value (ex: $product->setDescription()), as it has multiple conceptual and
 technical flaws (attribute codes are data, not a freeze structure, needed to maintain an full attribute cache in product
 that made the entity too smart for its own good and created performances problem)
 - Remove Product::createValue(). Can be replaced by calling ProductManager::createFlexibleValue() and setting attribute, scope and locale on the created value.
+- Product datagrid, hydrate rows as arrays (in place of objects) to reduce the loading time
+- Datagrid configuration, replace [flexible_entity] config by a [source][is_flexible] config to avoid to define the used entity twice
+- Rename and move src/Pim/Bundle/EnrichBundle/Resources/views/Completeness/_datagridCompleteness.html.twig => DataGridBundle/Resources/views/Property/completeness.html.twig
 
 # 1.0.0-rc-1 - "Tortoise Beats Hare" (2014-02-06)